+++ packages/axios-extension/src/auth/connection.ts
@@ -36,7 +36,7 @@ export class Cookies {
      */
     public addCookie(cookieString: string): Cookies {
         const cookie = cookieString.split(';');
-        const [key, ...values] = cookie[0]?.split('=');
+        const [key, ...values] = cookie[0]?.split('=') || [];
         const value = values?.join('='); // Account for embedded '=' in the value
         if (key && cookieString.indexOf('Max-Age=0') >= 0) {
             delete this.cookies[key];